// src/app/services/generated/api-client/model/apiResponsePagedResultProcessHistoryDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { PagedResultProcessHistoryDto } from './pagedResultProcessHistoryDto';


export interface ApiResponsePagedResultProcessHistoryDto { 
    body?: PagedResultProcessHistoryDto;
    requestID?: string;
    timeStamp?: string;
    status?: string;
    message?: string;
    i18nErrorCode?: string;
}