// src/app/services/generated/api-client/model/pagedRequestVoid.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface PagedRequestVoid { 
    size?: number;
    page?: number;
    sort?: PagedRequestVoid.SortEnum;
    filter?: object;
}
export namespace PagedRequestVoid {
    export const SortEnum = {
        Asc: 'ASC',
        Desc: 'DESC'
    } as const;
    export type SortEnum = typeof SortEnum[keyof typeof SortEnum];
}